+++ clients/ts-sdk/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.4.0] clients/ts-sdk
 ### Releases
 - [NPM v0.4.0] https://www.npmjs.com/package/patronts/v/0.4.0 - clients/ts-sdk
+
+## 2026-09-05 01:42:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.11.0] clients/ts-sdk
+### Releases
+- [NPM v0.11.0] https://www.npmjs.com/package/patronts/v/0.11.0 - clients/ts-sdk